BUG28303209 ----------- Also restore bug reported in BUG#90609. When the following conditions are true we will crash on a segfault in restore of an LCP. 1) More than 2002 parts in LCP to restore 2) Two LCP ctl files are found in recovery 3) The LCP ctl file to use is the first one
This can easily happen after long time of operating a 7.6 cluster and can happen
in many nodes at the same time.

Problem is that the m_lcp_ctl_file_data array is big enough to read the data in
from disk, but not big enough to also decompress the partPairs array from 24 bits
per part to 32 bits per part. This requires an additional 180 bytes.

Fixed by extending the size by the size of the LCP ctl header to ensure that the
parts array have at least 2048 32-bit words to its access.