WL#8693:Remove the syntax for GROUP BY ASC and DESC
In MySQL, historically GROUP BY has been used for sorting.
If a query specifies GROUP BY, output rows were sorted according
to the GROUP BY columns as if the query had an ORDER BY for the
same columns. This was in MySQL version 5.7 and earlier. However
it changed in MySQL 8.0 when implicit sorting for GROUP BY was
removed as part of WL#1074. But what still remained was the
non-standard extension that MySQL provided with GROUP BY clause
to specify the order for GROUP BY columns - GROUP BY column ASC/DESC.

This worklog removes the syntax for GROUP BY ASC and DESC.
With this and the earlier removal of implicit sorting for GROUP BY, code
that presumed that GROUP BY always sorts is removed too.