Bug#28142250: WRONG INITIALIZATION OF PTHREAD_RWLOCK IN LOG SERVICE
At start-up, we try to set up a basic logging pipeline very early.
The function that parses the configuration string for this first
asked the component framework if a service listed in the configuration
was known. If it was not, it would be compared to the names of built-in
services; if it did not match any of those either, it would fail.
When queried, the component framework would try to acquire its own
lock. This lock however would not have been initialized at that point,
as the full initialization (and indeed the loading of any external
components) would happen later in the start-up.

There is no pressing reason we should query the component framework
at that point in the start-up; on one hand, we're activating a
default pipeline from a compiled-in configuration string that we
know does not contain external components; on the other, external
components are not yet available at that point anyway.

This patch changes the order of tests in the configuration parser
so that we first check whether a log-service is internal, and query
the component framework only when that fails. As during start-up
we have a fixed configuration that consists solely of internal,
that is to say, known, services, the code path that queries the
component framework (where the uninitialized lock would be acquired)
is never used.