Bug#28311801: THE GLOBAL VALUE OF SECONDARY_ENGINE_EXECUTION_COUNT IS ALWAYS ZERO
Only the session value of the Secondary_engine_execution_count status
variable is updated when a statement is executed on a secondary
storage engine. The global value is not updated.

Fix it by updating LAST_STATUS_VAR to point to the
secondary_engine_execution_count variable, which is after the variable
currently pointed to by LAST_STATUS_VAR.

Change-Id: I22312a4835dd42da87d0641748e7a417cab5b4f9