Bug#28341794 CLEAN UP THE CONTENTS OF THE .GITIGNORE FILE
There are more than 2700 lines in the .gitignore file, most of them related
to in-source compilation (with autotools).

This is confusing and creates problems for adding legitimate file names to
the repo.

Clean up the file, leaving only editor-related stuff.
Also make cmake bail out if we do in-source builds.

Change-Id: I88896a8244affbd48033e347e60062fe7312dea3