Bug#27849293 INVALID WRITE IN SYSSCHEMA.PR_DIAGNOSTICS
In the case of a UNION where there are different character sets on each side of
the union, the optimizer might insert a CONVERT in order to get a matching
character set/collation on both sides. For example:

  SELECT 'foo' UNION ALL SELECT CONVERT('bar' USING utf8mb3);

...might be rewritten to

  SELECT CONVERT('foo' USING utf8mb3) UNION ALL SELECT CONVERT('bar' USING utf8mb3);

If this is in a prepared statement, the rewrite happens during the _execution_ of
the prepared statement, and not during prepare. So when altering the item tree,
the optimizer adds a "change record" that remembers what change was done, and
where it was done so it can be rolled back to its original state when the execution
is complete. In this particular case, the pointer to *where* the change happened
points to a stack-allocated array that is out of scope when the rollback
happens, and thus giving an invalid write with valgrind. This patch allocates the
array that contains the Item to be rolled back on a MEM_ROOT instead of on the stack.

Change-Id: I798e8986ab954f49952ae38490224f1bcdcda938