BUG#28341843: Fix restore issue with partial LCP
During restore of an LCP we can encounter a row that we want to insert that
is already present. In this case we need to delete it by PK rather than by
ROWID followed by an insert. This requires us to delete a row without
freeing the disk page entry for disk pages.