Bug#20597049 : MAIN.PARTITION_PRUNING FAILS OCCASIONALLY DUE TO DIFFERENCE IN EXPLAIN OUTPUT
Issue:
------
main.partition_pruning fails sporadically on PB2 with result content mismatch
in the filtered column of the explain output.

Fix:
----
The test was stabilized by using ANALYZE TABLE at required places within the test.
During testing it was observed that when --mysqld=--innodb-page-size=4k was
passed as an option to run the test, it failed with result content mismatch
in the type column of the explain output. To avoid such failures --source
include/have_innodb_16k.inc was added.