Bug#28238072 - BACKPORT BUG#27918666 'REMOVE THE EXIT() CALL FROM DIE_OP IN XCOM' TO 5.7
Description
-----------
When an XCom node becomes isolated from the group (e.g., due to a network
partition) it will lag behind the other members, missing the messages that
they exchange. When the connection to the group is re-established, it is
possible that all other nodes in the group have already garbage-collected
the old messages that the isolated node needs to get back into the group
without issuing a formal leave and join. When this happens, one of the
members of the group will send the node a 'die_op' so that it terminates
its execution. Currently, that termination is accomplished by means of an
exit() call, which causes the server to terminate abruptly. At the moment,
this is not a problem, since GCS expels nodes after 5 seconds (hence
preventing them from lagging behind for long enough for a 'die_op' to
actually be sent). However, after WL#11570 (in which the users will be
able to configure a delayed expel timeout) is pushed, the nodes could see
themselves in this situation.

Solution
--------
Backport the patch that has already been pushed to mysql 8.0, which
replaces the exit() call with a controlled termination of both XCom
and GCS: when XCom receives the die_op, it issues its own termination
and informs GCS about it; GCS then proceeds by terminating itself, in
the process installing a leave view to notify the application of the
unexpected termination. This backport also includes a fix for a
concurrency issue found in the original patch, which has since been
fixed in mysql 8.0 by WL#11570.