Bug #28317172 ASSERTION FAILURE: BTR0CUR.CC.*RW_LOCK_OWN(DICT_INDEX_GET_LOCK(INDEX)
Problem:

The flag BTR_ALREADY_S_LATCHED is supposed to be used only to determine if the
index lock has been taken in mode.  But its usage in
row_ins_scan_sec_index_for_duplicate() is such that it can refer to either S or
SX mode. So the assert needs to be relaxed to check either for S or SX mode.

Solution:

Relax the assert to check for either S or SX mode.

rb#20145 approved by Jimmy.