WL#12231 RUN AUTOTESTS WITH LATEST SCRIPT AND ATRT
Provides an extra option of running autotest tests with a version of
atrt which is not build-specific.

The autotest tests are currently running with build specific scripts
and atrt. This worklog provides an extra option of running autotest
with a different atrt binary preferrably the latest one by adding an
extra option called custom_atrt to autotest-run.sh. Another option
called custom-cpcc is also added to ensure that autotest tests can be
run with the recent version of ndb_cpcc.