WL#12271 ATRT REPORT DATABASE CONNECTION FAILURE
Modify atrt such that a warning message with the connection failure message
is issued whenever there is a failure in establishing a connection to MySQL
instance. Also modify the error message appropriately to avoid repetition
of connection failure messages.

A warning message describing the connection failure is printed before
printing each connection retry message to give a clear indication of why
the connection retry is happening. Also, the error message printed no
longer contains the error message but contains the host, port and socket
for which connection is failing.