WL#12229 ATRT SUPPORTS TEST FAILURE RETRIES
Adds support for ATRT to re-run the test cases in case of failure.

Modifies ATRT to accept a configuration file where user can specifies the number
of retries if a test case fails. In case of failure test case, all the processes
will be restarted and the test will retried up to the no of times as specified in
config file. Setting a retry value in config file is optional, by default retry value
is '0' , where as the retry value less than zero will be treated as an error and retry
value greater than five, issues an warning.

Updated the storage/ndb/test/run-test/README general algorithm to include retry
algorithm.

changes made in storage/ndb/test/run-test/check-tests.sh to support configuration
if "max-retries" parameter is specified.