WL9508 - Post-Fix Several tests fail periodically in PB2. They almost always have to do with waiting too long for the purge thread to truncate an inactive_explicit undo space and make it empty.
The function innodb_alter_undo_tablespace_inactive() is refactored here
to be more directed in what it does. In addition, the rule that the second
undo space can be made inactive is now changed to the third.  In other words,
there must always be 2 active undo spaces.  The third one may be explicitly
set empty. The reason is that while the third is being set inactive, one of
the other two can be implicitly truncated.  That would leave one more undo
space to use for new transactions.