Bug#26322656: ZERO TPS RECURS BY ACTIVE PURGETHREAD(UNDOLOGTRUNCATE=ON);     OFF SHOWS STEADY TPS
Instead of doing two checkpoints during the process of undo tablespace truncation,
just flush the dirty pages for that file from disk.

This change was committed to mysql-trunk for 8.0.14 after review in RB17522
Extracted from the full WL9508 patch and put into mysql-8.0 for 8.0.13 in this commit.
Tested and confirmed to eliminate the Zero TPS by Vince Rezula on 2018-07-27.