Bug#27798708 ALTER TABLE CRASHES: ASSERTION FAILURE: DICT0DICT.CC:1420: UT_A(TABLE2 == NULL)
PROBLEM:
when the create table with multi partitions fails after creating few partitions.
It leaves IBD files of partitions in the data directory and in SE.
This can have an adverse effect later.

Analysis:
SE is not doing the proper cleanup after a failed partition.

Fix:
call row_drop_database_for_mysql as part of the cleanup to remove the partition.

RB: 20055
Reviewed by: Jimmy Yang <Jimmy.Yang@oracle.com>