Bug#28212734 UPGRADE FAIL: ASSERTION FAILURE: DICT0DICT.CC:1272:TABLE->REFERENCED_SET.EMPTY()
Problem:
The server crashed while performing an in-place upgrade. The crash
happened while attempting to evict a table from the cache to make
some space to accomodate more tables. The table being evicted
had active foreign key relationship with some other table, causing
the server to hit an assert to assure that there are no active
foreign key relationships on this table.

The table being evicted had a FTS index. During upgrade, all the tables
with FTS index are added to a separate list. At the end of the upgrade,
all the tables in this list are marked as ready for eviction without
checking if they have any foreign key relationship. This results in moving
a table with foreign key relationship in the LRU list. The server
idle thread invokes and attempts to make some space in the cache by
evicting tables in the LRU chain and hits an assert while evicting the
table with FTS index incorrectly placed in the LRU chain while doing
the post-upgrade cleanup.

Solution:
Prevent eviction of the table during post-upgrade cleanup by not moving it
to the LRU chain, if it has a foreign key relationship.

Reviewed by: Allen Lai (zhlai@oracle.com)

RB# 20323