BU#28431343 - BASEDIR.TEST FAILS AT X PLUGIN STARTUP
Description
===========
The test starts MYSQLD without config file that was generated by MTR. This causes
that X Plugin configuration variables are set to their default values.
In case of UNIX socket, the default is '/tmp/mysqlx.sock'.

The default UNIX socket file (or lock file) is accessed by multiple MYSQLD instances
and may influance this test.

Fix
===
Testing X Plugin is not the main goal of this test, thus it is going to be disabled
at server restarts.

RB: 20203
Reviewed-by: Tomasz Stepniak <tomasz.s.stepniak@oracle.com>