BUG#28156802 - MYSQL 8 SERVICE GOES DOWN ON EXECUTING ONE PROCEDURE.
mysqld goes down when we execute a stored procedure that involves
CREATE ... SELECT statement and the SELECT clause has a LIMIT/OFFSET
value which is fetched via a SP variable. This happens in 8.0.
Due to refactoring, the OFFSET/LIMIT item fields are not fixed.
The fix is to fix the fields just before OFFSET/LIMIT sp variables
are evaluated as in 5.7.