Bug#26280041 RPL.RPL_HEARTBEAT FAILS SPORADICALLY AT CHANGE MASTER IN PB2
Scenario:

On the master, it "commits" three transactions (GTID 1, 2 and 3).

It replicate it to the slave during the first part of the test case.

Then, in the second part, a RESET SLAVE + START SLAVE is issued. From the
GTID point of view, without applying any of the received transactions, the
SQL thread is synced with the master, because the slave has already applied
the 1, 2 and 3 GTIDs. The slave fills relay log with transactions it already
applied, and the applier will apply them, but GTID auto skip will take care
of not actually executing the statements.

Next, a CHANGE MASTER specifying a host name will make the I/O thread to
start replicating from the first GTID again.

Finally, the test case wait for the SQL thread to be synced (it was
regardless of applying from relay log), and wait until SQL thread is waiting
from I/O thread to queue more events. A slow I/O thread might make SQL thread
enter the expected state without being in sync with master.

So, stopping the slave after the two checks would still lead to a
non-deterministic behavior (it may left gaps, so MTS must be recovered before
any CHANGE MASTER take place).

Fix:
A simple DROP on the master before synchronizing the slave SQL
thread made test case behavior more deterministic and test uses
MASTER_AUTO_POSITION now when GTID is enabled.