Bug#28199372 : NPE FROM CLUSTERJ WHEN "GET ALL" ON A TABLE CONTAINING TEXT OR BLOB FIELD
ClusterJ throws NullPointerException when a full table scan is
performed on a table with Blob or Text fields.

Any transaction involving the blob or text fields should initialise the
blob fields using activateBlobs function before actually executing the
transaction. The activateBlobs function internally calls the
getBlobHandle function on the respective ndbOperation and stores the
blob handle for reading later after transaction execution. In this
particular scenario involving full table scan, the activateBlobs or any
similar function is never called on the ndbOperation and the blob handles
never gets stored. Due to this when ClusterJ attempts to read the value
of the blob later, after the transaction execution, it throws a
NullPointerException. This patch fixes that by updating the
NdbRecordTableScanOperationImpl class to call activateBlobs when the
operation is being defined.

This patch also updates the QueryBlobIndexScanTest. The test previously
only tests index scans to retrieve blobs. Now it is updated to test also
table scans. For this purpose, a new column without any index is added
to the `blobtypes` table and also the related model is updated. The
test name is also updated to QueryBlobTypesTest reflecting these
changes.