Bug#28051645 : CLUSTERJ EXCEPTION "THE METHOD IS NOT VALID IN CURRENT BLOB STATE"
If a table contains a Blob or a Text field, and the table is queried
using ClusterJ for a record that doesn't exist, the exception "The
method is not valid in current blob state" is thrown.

The postExecuteCallback method defined in the NdbRecordKeyOperationImpl
attempts to read the blob values everytime after a query execution
without checking if the operation to fetch the key returned any errors.
Due to this, the above exception is thrown when there is no record to
read the blob from. This patch fixes that by updating the function to
read the blob values only when there is no error in executing the
ndbOperation. A very similar issue in the NdbRecordUniqueKeyOperationImpl
is also fixed in this patch. Also updated testcases.