Bug #28387760 BUILD FAILURE WITH -WERROR ENABLED USING GCC-8
Follow up patch eliminating the need for setting the 'Wno-ignored-qualifiers'
and '-Wno-cast-function-type' comiler flags.

Root cause for requiring the '-Wno-cast-function-type' option was the implicit_cast between
incompatible function pointers performed by  CachedArrayPool c'tor:

-  explicit CachedArrayPool(CallBack* seizeErrorHandler=NULL)
-     : ArrayPool<T>(reinterpret_cast<typename ArrayPool<T>::CallBack*>(seizeErrorHandler)) { }

Here we casted a pointer to a function taking a 'CachedArrayPoll<T>&' as its argument,
to a function taking an 'ArrayPoll<T>&' argument. Even if class CachedArrayPoll
is a subclass of ArrayPoll, and casting between these classes is allowed, the same is
NOT true for pointer to functions taking these classes as arguments. Thus
Gcc will throw a warning for the above cast.

The fix is to instantiate an abstract class ErrorHandler with different
implementation for the different 'seizeErrorHandler' functions required.
A virtual 'failure()' function is defined in this class which call the
provided  'seizeErrorHandler' functions upon a seize-failure.