Bug #28295814 INNODB.INNODB_WL6915_DEBUG OCCASIONALLY TIMESOUT ON PB2
Problem:

The crash point ib_trx_crash_during_xa_prepare_step is placed after
the call to trx_prepare_low().  This means that the trx has already
prepared as far as the file system is concerned.  Because of this
the test case innodb.innodb_wl6915_debug was failing due to timeout
(MDL locking issues).

Analysis:

Before the crash happened the trx was in TRX_STATE_ACTIVE, but the associated
undo log segment was set to TRX_UNDO_PREPARED in trx_prepare_low().  If
trx_prepare_low() is completed for the regular undo log segment of a trx, then
the trx is considered to be in prepared state during recovery.

If the intention of the crash point ib_trx_crash_during_xa_prepare_step is to
kill process before the trx is prepared, then it needs to be done before
trx_prepare_low().

Solution:

Move the crash point ib_trx_crash_during_xa_prepare_step before the
call to trx_prepare_low().

approved by Deb over IM.