WL#12064: Backport WL#11568 to MySQL 5.7
This worklog is created for the backport of the WL#11568: "Group
Replication:
option to shutdown server when dropping out of the group" to MySQL 5.7

On the backport to 5.7, we will keep the 5.7 existing behaviour, that
is,
when a failure happens the server switches to read_only mode.
The DBA will be able to choose the new behaviour by adjusting the option
group_replication_exit_state_action.