Bug 28262259: TRIGGERS ON SAME RECORD FIRED FOREVER -> ERROR: 233 OUT OF OPERATION RECORDS
  There are two variants of deferred trigger/constraints:

 1) Triggers created by a 'NO ACTION' foreign key are deferred by
    declaration, and managed by deferred<Op>Triggers list.
    These are always fired at commit time by
    Dbtup::checkDeferredTriggers() calling ::fireDeferredTriggers()

 2) In addition any constraints defined to be 'immediate' may be
    deferred by setting the 'ndb_deferred_constraints' or
   'TupKeyReq::deferred_constraints' flag in the TUPKEYREQ signal.

In either case, when TUP find that some triggers should be executed as deferred,
it informs TC about these outstanding triggers in the FIRE_TRIG_CONF and LQHKEYCONF
replies. At commit time TC will then send a FIRE_TRIG_REQ to TUP which is intended
to start the deferred trigger execution.

However, it turned out that deferred trigger firing did not distinguish between
whether it should fire only the triggers deferred by declaration (1), or if
the immediate triggers had been deferred. So it unconditionally re-fired the
immediate triggers, even if they already had been executed.

If the immediate triggers contained a self referencing cascading update constraint,
the update was then effectively re-executed in an endless loop. For other flavours
of immediate triggers, it was probably only an extra overhead firing the trigger twice.

This patch introduce a ::deferred_constraint flag in the Dbtup::Operationrec
object. This object is persistent across the initial LQHKEYREQ, and later FIRE_TRIG_REQ
which may fire and deferred triggers. It keeps track of whether the initial
LQHKEYREQ requested the immediate triggers to be deferred. The later FIRE_TRIG_REQ
will the conditionally fire the immediate trigger only when they has been deferred.