Bug#28451957 TRIGGER EXECUTION MAY REFERS A DROPPED TRIGGER DEFINITION
Trigger execution may use ::getPtr() to refer a trigger definition which has
already been dropped.

Root cause was a misconseption that ::getPtr() would return NULL if get'ting
an object already released, or accessing the ArrayPool outside of
its bounds.

However, the actuall behavior if ::getPtr() is:

- To just crash if accessed outside the bounds.
- Return the object, even if released, if within bounds (Crash in debug)

This patch replace the getPtr() where we may access a released trigger definition
with ::getPtrIgnoreAlloc(), which will eliminate the crash in debug compiled code.

The check for determining if the trigger definition record has been released
is changed to check for the ::triggerId to be as expected.

Dropping of the trigger definition will set ::triggerId to the invalid  0xfffff...
to indicate the the record is released.

(cherry picked from commit f4a0de5e6305adb4be25e45c5ef1d7bb02c52ad0)