BUG#28438281: MYSQL-TEST-RUN DOESN'T LOOK FOR MYSQL_CLIENT_TEST IN THE               CLIENT-BINDIR DIRECTORY
Description:
------------
For every other required binary, MTR checks that they exist in the
variable set with the '--client-bindir' option, but not for
'mysql_client_test' executable, causing test failures if it can't be
found.

Fix:
----
Added the location specified with '--client-bindir' option to the list
of locations to be searched for 'mysql_client_test' executable.

Change-Id: Iad26fdb84f79d37a4df45ff3bc526962f78fc510