Bug#28165060 MYSQL IS TRYING TO PERFORM A CONSISTENT READ BUT THE READ VIEW IS NOT ASSIGNED!
The problem is that when opening a heap table just after a server restart,
the open_table() attempts to write a binlog to delete all the heap
table rows at the slave. This binlog write calls commit operation. And this
commit is not expected to be invoked when fetching I_S table statistics.
This causes an assert in InnoDB while processing SELECT query.

This patch enables open_table_entry_fini() to ignore handling
implicit_emptied property (which is only for heap tables) when I_S query is
opening this table to read table statistics. I.e., we avoid writing binlog
to delete all heap table rows, which intern avoids calling commit.

A test case is added in main.information_schema_statistics.

Change-Id: I5299a5c2882c90036a26947107019b29f7e93eca