BUG#27539838: NOT ALL ABORTED CONNECTS ARE REPORTED TO               ERROR.LOG PROPERLY
ANALYSIS:
=========
Unsuccessful connection attempts are not being logged in
the error log when log_error_verbosity=3.

For every unsuccessful connection attempt, the status
variable 'Aborted_connects', which keeps the count of
failed attempts to connect to the MySQL server, is being
incremented. But the error log contains no information
reflecting that.

FIX:
====
We are now logging appropriate information in the error log
for every unsuccessful connection attempt if
log_error_verbosity=3.

Change-Id: Ie8e6496d4f81f728db0afbd420ff84f3bcce79fa