Bug#28499603 SHOW INDEXES PRINTS WRONG OUTPUT FOR PREFIX INDEXES
When executing "SHOW INDEXES FROM tbl;", the server gives the client
a list with information about all key parts in the given table. If the
key part is a prefix index (only a subpart of the column is indexed),
the prefix index length will be given in the column "Sub_part". If the entire
column is indexed, the "Sub_part" will be set to SQL NULL.

After WL#6599 was implemented, this column started to return SQL NULL if the
prefix index length was 8 bytes for a BLOB/TEXT column, indicating that the entire
column is indexed. This is wrong, and the reason is that we started to compare
the prefix index length against the column length instead of the column _key_
length. And, for legacy reasons, the server sets the column length for BLOB/TEXT
columns to 8 regardless of the actual length. This patch changes the server back to
comparing the prefix index length and the column _key_ length in order to
determine whether or not the entire column is indexed.

Note that this only affects output from SHOW INDEXES and INFORMATION_SCHEMA.STATISTICS,
and not any DML/DDL.

Change-Id: Ie1d0a1fb78cee7629569394d2b01f9bf1d2b389a