Bug#28068224: CONFIG ISSUE MAKES SERVER SKIPS OWN DETAILS IN ER_GRP_RPL_MEMBER_CHANGE MSG
Problem: COMMA GETS ADDED FOR OFFLINE MEMBERS

Description:
While displaying active members of the group during view change only ONLINE or
RECOVERING members are considered for display.
Members are separated by COMMA. When member is not ONLINE or RECOVERING member
is not considered to display as group member but COMMA still gets added to the
string output.

Resolution:
Removed condition to display information of all members in error log message.