Bug #27921932   UPDATING ROW WITH GENERATED COLUMN DURING ALTER TABLE RESULTS IN BROCKEN DATA
PROBLEM

While an online ALTER and UPDATE are going in parallel, the Update log
applied by ALTER to the new table doesn't take into account the computed
value of generated column in the old row while we update the secondary
index. This causes a segfault when another update is done on table as
row not found error is returned.

FIX

Make sure the computed value of virtual column is updated to old row
that is used to do update to secondary index.

Reviewed by: Jimmy Yang <Jimmy.Yang@oracle.com>
RB: 20339