BUG#28378971 - GR:UNREACHABLE NODE CRASHES WHEN IT TRIES TO JOIN THE GROUP BEFORE EVICTION.
Description
-----------
If a server becomes unreachable (e.g., due to a network loss) for some
time while the remaining servers operate under high load, when it becomes
reachable again it will receive a large message backlog to process very
quikly. This will cause the XCom cache to become exhausted and XCOM will
not be able to obtain new pax machines, something for which it is not
prepared.

Solution
--------
Modify XCom so that when it is not possible to obtain new cache entries the
task that is requesting the entry waits (with a timeout) until one becomes
available.