WL#9756: Fixup: do_virtual_alloc
do_virtual_memory is used in debug build to emulate huge amount of memory,
making page numbers as high as possible.

This is done by reserving a huge amount of virtual memory but only
populate part of them.

Problem
=======
For Windows and Linux the virtual memory space that is not populate are
marked so that they will not be included in an eventual core dump.

For other OS like for example MacOSX no such support is implemented and
huge core dumps will follow when core dumping.

Solution
========
This patch only enables do_virtual_alloc for Windows and Linux there we
support not dumping the whole virtual memory.