Bug#28486390 False possitive 'Key constraint violations' reported
The root cause for this bug has been tracked down to how the
'savePointId' is set for the parent/child triggers.

The savePointId is an internal mechanism controlled by TC which
decide which 'version' of a tuple which should be seen if the
tuple is modified (possible multiple times) within a transaction.
For an 'immediate trigger' the parent/child triggers should see
the tuple contents as immediately *after* the operation which updated
it. This was implemented by setting the triggers LqhKeyReq::savePointId
from ApiConnectRecord::currSavePointId. This relied in that
ApiConnectRecord::currSavePointId was incremented when TC executed
the 'last' of the operations sent from API.

However, TC could start generating the LqhKeyRequests for the
row-triggers at the receipt of LQHKEYCONF,  before it had
executed the 'last' operation. Thus they were produced with an
incorrect savePointId, effectively resulting in the 'before'
image value of the tuple being used when validating the parent/child
relations. (Effectively we had implemented an 'intermediate' trigger
instead of an 'immediate (after)' trigger'.

The handling of deferred triggers was not affected by this bug.
These triggers will be fired as part of the commit processing,
where all preceeding *KEYREQ's has completed, and the
ApiConnectRecord::currSavePointId correctly contains the
last savePointId in the transaction.

This patch fix how the LqhKeyReq::savePointId is set in the
parent/child *read* triggers: If the trigger is an immediate trigger,
the savePointId is set a '+1' relative to the originating update
operation. This ensures that the read operation verifying the parent/child
consistency (only) sees the updated 'after' value.

Patch also changes a few places where we checked for the apiConnectstates
CS_SEND_FIRE_TRIG_REQ or CS_WAIT_FIRE_TRIG_REQ. This is what the method
isExecutingDeferredTriggers() is intended to do, so rewrote to call this
method instead.

The unused '#define INTERNAL_TRIGGER_TCKEYREQ_JBA' was also removed.