Bug#28500861 DBTUP crash due to 'Tuple corruption detected'
When TUP abort an INSERT operation, it will change the 'm_header_bits'
for the tuple from 'ALLOC' to 'FREE'. The FREE tuple is also
marked as invalid by calling Dbtup::setInvalidChecksum().

If the same transaction also contained a READ of the same tuple,
there could be a race between ACC granting access (ACCKEYCONF),
and the ABORT (TC -> LQH -> TUP) undoing the INSERT by setting
the tuple to FREE. Thus the READ may be resumed and find the tuple FREE.
This effectively caused a corrupted tuple to be READ, thus the crash.

This patch fix Dbtup::setup_read() to detect a FREE tuple and
return ZTUPLE_DELETED_ERROR if detected.