Bug #28525898 NDB_RESTORE MEMORY LEAK WHEN TABLE ALREADY EXIST
The ndb_restore_misc and ndb_restore_compat_compression tests fails when run
with ASAN in mysql-trunk. Memory allocated for BackupRestore instance in
readArguments() is freed by the function clearConsumers(). But, this function
is not called from all error paths. So, memory is leaked.

Fixed the issue by calling clearConsumers() in all error paths.