BUG#28372628: Fix LCP_SKIP issue
When performing a node restart we perform a copy from live node
to starting node of changed rows. This copy process also involves
sending deleted rowids to the starting node.

In a node an update, delete or insert of a row sets the GCI of the
row to the GCI it was written in. There is one exception to this
rule, this is during node restart and copy from live node to
starting nodes for deleted rows. These rows can at times have
GCI equal to 0. This GCI is then copied over to the starting node.

If a row existed in this position AND a full local LCP was started
due to UNDO log getting full, we could set the LCP_SKIP bit on a
row with GCI = 0.