Bug#28531148 - ASSERTION FAILURE: ROW0MYSQL.CC:3914 DURING DROP TABLE
The assertion says that the table has been referenced by another thread
when it is being dropped. So there must be some MDL problem on the table.
Actually another thread is still accessing this table via memcached API,
it just released the MDL but yet to release the table itself.
This is incorrect because MDL should protect the whole period of accessing
the table

So the fix should be in memcached API ib_cursor_close(), it should first
release the table, and then release the MDL.

RB: 20360
Allen Lai <zheng.lai@oracle.com>