WL#12114: MTR SUPPORT FOR AUTOMATIC LOADING OF DATA TO SECONDARY           ENGINE AFTER A DDL STATEMENT
Description:
------------
This WL will introduce a new MTR option --change-propagation. This
option must be used along with --secondary-engine=engine_name option.

When this option is defined or enabled, MTR will automatically change
the secondary engine of a table to 'engine_name' and will load the
table contents from primary engine to secondary engine after a DDL
statement.

The scope of this WL is limited to following DDL statements

  1. ALTER TABLE (ALTER TABLE ... RENAME newtab not supported)
  2. CREATE TABLE
  3. CREATE INDEX
  4. DROP INDEX

Change-Id: Ic5e72f5eba6350737a56165ee72ba2fe9e9f779a